organized the directory files.
